--- src/main/java/roomescape/HomeController.java
@@ -1,12 +1,0 @@
-package roomescape;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-    @GetMapping("/")
-    public String home() {
-        return "home";
-    }
-}
