--- src/main/java/roomescape/Reservation.java
@@ -1,4 +1,0 @@
-package roomescape;
-
-public record Reservation(int id, String name, String date, String time) {
-}
