+++ src/main/java/roomescape/RoomescapeApplication.java
@@ -10,3 +10,4 @@ public class RoomescapeApplication {
     }
 
 }
+
