+++ src/main/java/roomescape/dto/Reservation.java
@@ -1,0 +1,4 @@
+package roomescape.dto;
+
+public record Reservation(int id, String name, String date, String time) {
+}
